requirements-colab.txt
# PyNucleus-Model: Google Colab Requirements
# Minimal dependencies for running PyNucleus in Google Colab

# Core ML dependencies (Colab-optimized versions)
torch>=2.0.0,<2.4.0
transformers>=4.30.0,<4.50.0
sentence-transformers>=2.2.0,<2.8.0
langchain>=0.2.0,<0.4.0

# Vector store - CPU only for Colab
faiss-cpu>=1.7.0,<1.9.0

# Essential utilities
python-dotenv>=1.0.0,<1.1.0
pydantic>=2.0.0,<2.10.0
requests>=2.31.0,<2.33.0
beautifulsoup4>=4.12.0,<4.13.0
tqdm>=4.66.0,<4.67.0
numpy>=1.24.0,<1.27.0
pandas>=2.0.0,<2.3.0

# Document processing
pypdf>=4.0.0,<4.4.0
python-docx>=1.1.0,<1.2.0

# Configuration and validation
jsonschema>=4.0.0,<4.24.0
typing-extensions>=4.0.0,<4.13.0

# DSPy for structured prompting
dspy-ai>=2.4.0,<2.6.0

# HTTP clients
httpx>=0.24.0,<0.28.0

# Logging
PyYAML>=6.0.0,<6.1.0
structlog>=23.0.0,<24.5.0

# Enhanced embeddings
accelerate>=0.24.0,<0.35.0

# Optional document processing (lighter versions for Colab)
PyMuPDF>=1.23.0,<1.25.0
Pillow>=10.0.0,<10.5.0

# CLI framework
typer>=0.12.0,<0.13.0

# Note: Some dependencies like pythonnet (for DWSIM) are not compatible with Colab
# Use the mock DWSIM mode for testing in Colab environment 